Изменение классов с тестами валидтора
